+++ app/lab/murajaah-at-taisir/hooks/useMurajaah.ts
@@ -2,7 +2,6 @@ import { useState, useEffect } from "react";
 import useSWR from "swr";
 
 import { DEFAULT_MODE } from "../context/MurajaahContext";
-import { fetchAllSurah } from "../lib/fetchAllSurah";
 import { fetchAyahBySurahNAyah } from "../lib/fetchAyahBySurahNAyah";
 import { fetchSurahByNumber } from "../lib/fetchSurahByNumber";
 
@@ -10,13 +9,9 @@ export function useMurajaah() {
   const [selectedSurah, setSelectedSurah] = useState<number | null>(null);
   const [startAyah, setStartAyah] = useState<number | null>(null);
   const [endAyah, setEndAyah] = useState<number | null>(null);
+  const [randoming, setRandoming] = useState(false);
   const [randomAyah, setRandomAyah] = useState<number | null>(null);
   const [mode, setMode] = useState(DEFAULT_MODE);
-
-  const { data: allSurah, isLoading: isLoadingSurah } = useSWR(
-    "all-surah",
-    fetchAllSurah
-  );
 
   const { data: selectedSurahData } = useSWR(
     selectedSurah ? `surah-by-number-${selectedSurah}` : null,
@@ -41,9 +36,27 @@ export function useMurajaah() {
     }
   }, [selectedSurahData]);
 
+  const generateRandomAyah = () => {
+    if (!startAyah || !endAyah) return;
+    if (startAyah < 1 || endAyah > maxAyah) return;
+    if (startAyah > endAyah) return;
+
+    const randomedAyah =
+      Math.floor(Math.random() * (endAyah - startAyah + 1)) + startAyah;
+
+    if (startAyah === endAyah) {
+      setRandomAyah(randomedAyah);
+      return;
+    }
+
+    setRandoming(true);
+    setTimeout(() => {
+      setRandomAyah(randomedAyah);
+      setRandoming(false);
+    }, 500);
+  };
+
   return {
-    allSurah,
-    isLoadingSurah,
     selectedSurah,
     setSelectedSurah,
     startAyah,
@@ -52,6 +65,8 @@ export function useMurajaah() {
     setEndAyah,
     randomAyah,
     setRandomAyah,
+    randoming,
+    generateRandomAyah,
     mode,
     setMode,
     ayahData,